requirements.txt
# Flask e extensões
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0

# Banco de dados
PyMySQL==1.1.0
Werkzeug==3.0.1

# Utilitários
Pillow==10.1.0
python-dateutil==2.8.2

# Desenvolvimento e testes
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Produção
gunicorn==21.2.0
redis==5.0.1

# Cache e sessões
Flask-Caching==2.1.0
Flask-Session==0.5.0

# Segurança
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0

# Logging e monitoramento
structlog==23.2.0 